Use different output location if building with gradle.

+++ lib/high_five/thor/tasks/dist.rb
@@ -76,8 +76,13 @@ module HighFive
           android_name = HighFive::AndroidHelper.project_name_from_build_xml("#{android_path}/build.xml")
 
           if @output_file_name
-            say "copying final build #{android_path}/bin/#{android_name}-release.apk -> #{android_path}/bin/#{@output_file_name}.apk"
-            FileUtils.cp("#{android_path}/bin/#{android_name}-release.apk", "#{android_path}/bin/#{@output_file_name}.apk")
+            if File.exists?(gradle_file)
+              say "copying final build #{android_path}/build/outputs/apk/android-release.apk -> #{android_path}/build/outputs/apk/#{@output_file_name}.apk"
+              FileUtils.cp("#{android_path}/build/outputs/apk/android-release.apk", "#{android_path}/build/outputs/apk/#{@output_file_name}.apk")
+            else
+              say "copying final build #{android_path}/bin/#{android_name}-release.apk -> #{android_path}/bin/#{@output_file_name}.apk"
+              FileUtils.cp("#{android_path}/bin/#{android_name}-release.apk", "#{android_path}/bin/#{@output_file_name}.apk")
+            end
           end
         end
